Dev (#3)
* adding init_app()

This commit adds an init_app function which is provided by many flask
extensions to be more consistent with the flask application factory
format. (http://flask.pocoo.org/docs/1.0/patterns/appfactories/)

This function will also load the parameter for argon2 from the
flask application config.

* added exception handling

This commit adds the exceptions VerificationError and InvalidHash which
can be raced by argon2. Also a test-case for InvalidHash is added.

* updated readme